Add the option mangle: false, into production webpack.
The mangle option change the variables names, and that was causing a bug into my Skill's component.
My Skill's component uses the Component's name to execute or not a operation, so
if the component's name is changed in production, i have no way to now the name that will be used,
causing problems.